requirements.txt
click==6.7
cycler==0.10.0
matplotlib==2.0.0
numpy==1.13.1
pandas==0.20.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-dotenv==0.6.5
pytz==2017.2
six==1.10.0
